Validate `sample_weight` keys when passed as dict
Fixes that when passing `sample_weight` as a dict to .fit() with
the wrong dict keys keras would silently not perform any weighting.

If the same is done for the `x` or `y` arguments the backend will throw
an exception later, so there is no crucial need to update
`standardize_user_data` in the same way.